yt_searcher_mcp/requirements.txt
google-api-python-client>=2.60.0
openai>=1.3.0
serpapi>=0.1.5
beautifulsoup4>=4.10.0
requests>=2.28.0
gradio>=4.0.0
python-dotenv>=1.0.0
